few changes
- Buffs to Harper's heat resistances
- All gryphons slightly higher body size and hunger rate
- Adjusted armor ratings for all gryphons (some buffed, sharp nerfed)
- All Gryphons less likely to get food poison
- Sol and Harper spawn rarer (may rename to species name)
- Removed accidental compatibility with 1.4
- Beginning structure for 1.6

So because I will be renaming Sol and Harper to their species name 1.5 saves will not be compatible with 1.6 and you may need to start a new save or spawn them back in with a pawn editor. this also mean 1.5 will be the last version to eventually drop but all future ones I will likely leave for long term. (wont be likely to fix bugs in older versions however)